==> notebooks/1_csv_merging.ipynb <==
# Glob is used for matching pathnames for easily loading the 5 .csv files
import glob
import pandas as pd

### Creating a list of df for merging them.

# matching files that end in .csv in the data folder
path = r'../data/'
files = glob.glob(path + "/accidents*.csv")

all_dfs = [pd.read_csv(f) for f in files]

# Checking if columns are the same between all datasets
column_list = [df.columns.tolist() for df in all_dfs]

all(x == x for x in column_list)

### Columns are the same, we can merge the dfs.

merged_data = pd.concat(all_dfs, ignore_index=True)

# Checking if number of rows are the same for the merged and the separated dataframes
sum([df.shape[0] for df in all_dfs]) == merged_data.shape[0]

### Saving to a .csv file.

merged_data.to_csv('../data/merged_data.csv')
