==> notebooks/2_cleaning.ipynb <==
### Importing the merged .csv as a DataFrame.

import pandas as pd
import numpy as np
from datetime import datetime

pd.set_option('display.max_columns', 50)

df = pd.read_csv('../data/merged_data.csv', index_col=0)

### Checking columns names and dtypes.

df.info()

### Checking nan values.

df.isna().sum()

### Columns to be dropped.
- Accident_Index: Not required since we can simply use the DataFrame index.
- Location_Easting_OSGR and Location_Northing_OSGR: Those are GB-specific coordinates. We have Lat and Lon and therefore we do not need them.
- Police_Force: No info.
- Local_Authority_(District) and Local_Authority_(Highway): Not useful for our analysis.
- 1st_Road_Class and 1st_Road_Number: Only related to junctions. We already have many attributes that are useful for our analysis and prediction.
- Junction_Detail and Junction_Control: almost no data.
- 2nd_Road_Class and 2nd_Road_Number: Same as above.
- Pedestrian_Crossing-Human_Control: Very specific information, not useful for our analysis.
- Special_Conditions_at_Site: We already have a lot of information on conditions.
- LSOA_of_Accident_Location: Again, we have Lat and Lon.

columns_to_drop = ['Accident_Index', 'Location_Easting_OSGR', 'Location_Northing_OSGR', 'Police_Force', 
                   'Local_Authority_(District)', 'Local_Authority_(Highway)', '1st_Road_Class',
                   '1st_Road_Number', 'Junction_Detail', 'Junction_Control', '2nd_Road_Class',
                   '2nd_Road_Number', 'Pedestrian_Crossing-Human_Control', 'Special_Conditions_at_Site',
                   'LSOA_of_Accident_Location']

df.drop(columns_to_drop, axis=1, inplace=True)

df.info()

### Renaming columns.

column_names = {'Longitude': 'lon', 'Latitude': 'lat', 'Accident_Severity': 'severity', 
                'Number_of_Vehicles': 'num_vehicles', 'Number_of_Casualties': 'num_casualties',
                'Date': 'date', 'Day_of_Week': 'doy', 'Time': 'time', 'Road_Type': 'road_type',
                'Pedestrian_Crossing-Physical_Facilities': 'ped_crossing', 'Light_Conditions': 'light_cond',
                'Weather_Conditions': 'weather', 'Road_Surface_Conditions': 'road_cond',
                'Carriageway_Hazards': 'hazards', 'Urban_or_Rural_Area': 'urb_or_rur',
                'Did_Police_Officer_Attend_Scene_of_Accident': 'police_presence', 'Year': 'year'}

df.rename(columns=column_names, inplace=True)

df.info()

### Given the very low number of Nans we will be dropping all the rows containing them to make the analysis and prediction easier.

len(df.dropna())/len(df)

df.dropna(inplace=True)

### Converting date column to datetime.

df.date.head()

pd.to_datetime(df.date)

### We got no errors therefore the dates are all formatted in the same way.

df.date = pd.to_datetime(df.date)

### Removing minutes and keeping only the hours from Time column. Going to the next hour if minutes are > 30 and then changing all the 24s to 0.

# slicing the string to keep only the hours
func = lambda x: int(x[:-3]) if int(x[-2:]) <= 30 else int(x[:-3]) + 1
df.time = df.time.apply(func)

df.time.replace(24, 0, inplace=True)

### Checking .head again.

df.head()

### Checking the urban or rural column.

df.urb_or_rur.value_counts()

### Since 3 corresponds to undefined and it's only present in 35 observation we will be dropping the relevant rows.

df = df[df.urb_or_rur != 3]

### Types are fine and columns were correctly renamed. We will save the df to a new .csv file for the analysis and prediction.

df.to_csv('../data/clean_data.csv')
