==> notebooks/4_Preprocessing.ipynb <==
# Preprocessing
Preprocessing is the previous step to Machine Learning, it consists on preparing the data to optimize the model algorithm. We will follow these steps:
- **Get dummy variables:** categorical values need to be separed in boolean columns.
- **Rescaling:** normalization of the numerical values in order to not imbalance the algorithm.

## Get dummy variables

# importing necessary libraries
import pandas as pd
import numpy as np

# importing the data
df = pd.read_csv('../data/clean_data.csv', index_col=0)

# exploring which are the columns that need to be dummies:
df.head()

#### Police presence
This column, which tells us whether the police did attend the accident or not, has only two values: Yes or No. We will change these values to 0 and 1:
- 0 will mean no.
- 1 will mean yes.

# we will use the numpy where function
df['police_presence'] = np.where(df['police_presence'] == 'Yes', 1, 0)

#### Urban or rural
This columns is telling us if the accident happened in urban or rural.

# checking values
df['urb_or_rur'].value_counts()

We have two values: 1 for urban and 2 for rural. We will convert these into:
- 0 for urban.
- 1 for rural.

# we only need to subtract 1 from the column
df['urb_or_rur'] = df['urb_or_rur'] - 1

#### Hazards
The 'hazards' column tells us about risks and dangers the accident had.

# checking values
df['hazards'].value_counts()

We have 6 different hazards taking into account the 'none' one. These will be put in dummy variables, meaning that there'll be one column per hazard with a 1 in the rows with that hazard and a 0 in the ones without it. The hazard 'None' will be represented as a 0 in all the other columns.

# we will use the pandas get_dummies function
hazards = pd.get_dummies(df['hazards'], prefix='hazard')

# merging them
df1 = pd.merge(df, hazards, left_index=True, right_index=True)

# dropping the useless columns
df1.drop(['hazards', 'hazard_None'], axis=1, inplace=True)

#### Road condition
Tells us the conditions of the road, we have the following values:

df['road_cond'].value_counts()

We'll get dummy variables from here, as we did for hazards. Also we'll drop the 'Dry' value because it means it has none of the other and the conditions are normal.

# we will use the pandas get_dummies function
road_cond = pd.get_dummies(df['road_cond'], prefix='road')

# merging them
df2 = pd.merge(df1, road_cond, left_index=True, right_index=True)

# dropping the useless columns
df2.drop(['road_cond', 'road_Dry'], axis=1, inplace=True)

#### Weather
Tells us weather conditions , we have the following values:

# checking values
df['weather'].value_counts()

We'll get dummy variables from here, as we did for hazards. This time we won't drop values because we have the 'fine without high winds' as well as 'unknown' and 'other', which we may not exactly know what they mean but could give us information.

# we will use the pandas get_dummies function
weather = pd.get_dummies(df['weather'], prefix='weather')

# merging them
df3 = pd.merge(df2, weather, left_index=True, right_index=True)

# dropping the useless columns
df3.drop(['weather'], axis=1, inplace=True)

#### Light conditions
Tells us light conditions, we have the following values:

df['light_cond'].value_counts()

We'll get dummy variables from here, as we did for hazards. Also we'll drop the 'Daylight' value because it means it has none of the other and the conditions are normal.

# we will use the pandas get_dummies function
light = pd.get_dummies(df['light_cond'])

# merging them
df4 = pd.merge(df3, light, left_index=True, right_index=True)

# dropping the useless columns
df4.drop(['light_cond', 'Daylight: Street light present'], axis=1, inplace=True)

#### Pedestrian crossing
We have the following values:

df['ped_crossing'].value_counts()

We'll get dummy variables from here, as we did for hazards. Also we'll drop the 'no physical crossing...' value because it means it has none of the other and the conditions are normal.

# we will use the pandas get_dummies function
peds = pd.get_dummies(df['ped_crossing'])

# merging them
df5 = pd.merge(df4, peds, left_index=True, right_index=True)

# dropping the useless columns
df5.drop(['ped_crossing', 'No physical crossing within 50 meters'], axis=1, inplace=True)

#### Road types
We have the following values:

df['road_type'].value_counts()

We'll get dummy variables from here, as we did for hazards.

# we will use the pandas get_dummies function
rtypes = pd.get_dummies(df['road_type'], prefix='rtype')

# merging them
df6 = pd.merge(df5, rtypes, left_index=True, right_index=True)

# dropping the useless columns
df6.drop(['road_type'], axis=1, inplace=True)

#### Day of the week
We have the following values:

df['doy'].value_counts()

We will divide the days of the week in weekend (Fri, Sat and Sun) and rest of the week, (Mon, Tue, Wed, Thu).

# we will use np.where for this purpose
df6['weekend'] = np.where((df.doy == 6) | (df.doy == 7) | (df.doy == 1), 1, 0)

df6.weekend.isna()

#### Time
We have 24 values indicating the our of the accident, we want to convert them into day and night.
We'll create a column named night that will contain 0 if it's between 7 and 22 o'clock and 1 for the rest, considering it the night.

df6['night'] = np.where((df['time'] >= 7) & (df['time'] <= 22), 0, 1)
df6.drop(['time'], axis=1, inplace=True)

#### Now that we've finished converting dummy variables, we'll drop columns that don't give relevant information for our analysis.

df6.drop(['lat','lon','date', 'year'], axis=1, inplace=True)

df6.info()

## Rescaling 
Rescaling of the numerical values in order to not imbalance the algorithm.

# we've tried some rescaling methods from sklearn to get the data between 0 and 1, but none of them worked correctly
# we'll divide the speed limit by the max we have

df6['Speed_limit'] = df6['Speed_limit']/df6['Speed_limit'].max()

# checking everything looks okay
df6.head()

df6.to_csv('../data/preprocessed_data.csv')
